#1171 Reduce size of the Metalama.Tool package

+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.UnitTests/DesignTime/UserProfile/UserProfileTests.cs
@@ -2,7 +2,7 @@
 // SharpCrafters s.r.o. licenses this file to you under either the MIT license or a proprietary license, depending on the repository from which it was obtained.
 // Refer to LICENSE.md in the repository root for complete details.
 
-using Metalama.Framework.DesignTime.Diagnostics;
+using Metalama.Framework.ConfigurationFiles;
 using Microsoft.CodeAnalysis;
 using Newtonsoft.Json;
 using System.Collections.Immutable;
@@ -15,7 +15,7 @@ namespace Metalama.Framework.Tests.UnitTests.DesignTime.UserProfile
         [Fact]
         public void DiagnosticUserProfileSerialization()
         {
-            var originalDiagnostic = new UserDiagnosticRegistration( "MY001", DiagnosticSeverity.Error, "Category", "Title" );
+            var originalDiagnostic = new UserDiagnosticRegistration( "MY001", (int) DiagnosticSeverity.Error, "Category", "Title" );
 
             var file = new UserDiagnosticsConfiguration
             {
